Split shared types into separate file that can be imported from main, reload, or renderer
