+++ app/src/main/utils.ts
@@ -2,10 +2,3 @@
 export function sum(a: number, b: number): number {
   return a + b;
 }
-
-type JSONPrimitive = string | number | boolean | null;
-type JSONArray = JSONValue[];
-export interface JSONObject {
-  [key: string]: JSONValue;
-}
-export type JSONValue = JSONPrimitive | JSONArray | JSONObject;
