+++ app/src/preload/index.ts
@@ -1,7 +1,7 @@
 // See the Electron documentation for details on how to use preload scripts:
 // https://www.electronjs.org/docs/latest/tutorial/process-model#preload-scripts
 import { contextBridge, ipcRenderer } from "electron";
-import { ProxyRequest } from "../main/proxy";
+import type { ProxyRequest } from "../types";
 
 const electronAPI = {
   request: (request: ProxyRequest) => ipcRenderer.invoke("request", request),
